fix test suite: de-doctest code block
The code block was failing the test suite as a doc test. This was
happening, specifically, because of the method overwrite warnings.

None of that code block is appropriate for being a doctest, e.g., it
tests the exact form of the human-readable `show` output. So the fix is
easy.